final_commit

+++ src/Controller/DAController.php
@@ -15,21 +15,20 @@ use Symfony\Component\Routing\Attribute\Route;
 final class DAController extends AbstractController
 {
     #[Route(name: 'app_d_a_index', methods: ['GET'])]
-public function index(Request $request, DARepository $dARepository): Response
-{
-    $searchTerm = $request->query->get('search');
-    $monthDA = $request->query->get('month_da');
-    $monthBCA = $request->query->get('month_bca');
-    $retardDABCA = $request->query->get('retard_dabca');
+    public function index(Request $request, DARepository $dARepository): Response
+    {
+        $searchTerm = $request->query->get('search');
+        $monthDA = $request->query->get('month_da');
+        $monthBCA = $request->query->get('month_bca');
+        $retardDABCA = $request->query->get('retard_dabca');
 
-    $dAs = $dARepository->searchByFieldsAndMonths($searchTerm, $monthDA, $monthBCA, $retardDABCA);
+        $dAs = $dARepository->searchByFieldsAndMonths($searchTerm, $monthDA, $monthBCA, $retardDABCA);
 
-    return $this->render('da/index.html.twig', [
-        'd_as' => $dAs,
-        'searchTerm' => $searchTerm,
-    ]);
-}
-
+        return $this->render('da/index.html.twig', [
+            'd_as' => $dAs,
+            'searchTerm' => $searchTerm,
+        ]);
+    }
 
     #[Route('/new', name: 'app_d_a_new', methods: ['GET', 'POST'])]
     public function new(Request $request, EntityManagerInterface $entityManager): Response
@@ -53,16 +52,24 @@ public function index(Request $request, DARepository $dARepository): Response
     }
 
     #[Route('/{id}', name: 'app_d_a_show', methods: ['GET'])]
-    public function show(DA $dA): Response
+    public function show(int $id, DARepository $dARepository): Response
     {
+        $dA = $dARepository->find($id);
+
         return $this->render('da/show.html.twig', [
             'd_a' => $dA,
         ]);
     }
 
     #[Route('/{id}/edit', name: 'app_d_a_edit', methods: ['GET', 'POST'])]
-    public function edit(Request $request, DA $dA, EntityManagerInterface $entityManager): Response
+    public function edit(Request $request, int $id, DARepository $dARepository, EntityManagerInterface $entityManager): Response
     {
+        $dA = $dARepository->find($id);
+
+        if (!$dA) {
+            throw $this->createNotFoundException('Demande d\'achat introuvable.');
+        }
+
         $form = $this->createForm(DAForm::class, $dA);
         $form->handleRequest($request);
 
@@ -80,9 +87,15 @@ public function index(Request $request, DARepository $dARepository): Response
     }
 
     #[Route('/{id}', name: 'app_d_a_delete', methods: ['POST'])]
-    public function delete(Request $request, DA $dA, EntityManagerInterface $entityManager): Response
+    public function delete(Request $request, int $id, DARepository $dARepository, EntityManagerInterface $entityManager): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$dA->getId(), $request->getPayload()->getString('_token'))) {
+        $dA = $dARepository->find($id);
+
+        if (!$dA) {
+            throw $this->createNotFoundException('Demande d\'achat introuvable.');
+        }
+
+        if ($this->isCsrfTokenValid('delete' . $dA->getId(), $request->getPayload()->getString('_token'))) {
             $entityManager->remove($dA);
             $entityManager->flush();
         }
